LayoutTests/css3/calc/trig-functions-with-constants-expected.txt

Tests functionality of sin, cos, tan, e and pi

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



element.style["width"] = "calc(sin(pi/2) * 100px)"
PASS 100 is 100

element.style["width"] = "calc(cos(pi - 3.14159265358979323846) * 100px)"
PASS 100 is 100

element.style["width"] = "calc(cos(e - 2.7182818284590452354) * 100px)"
PASS 100 is 100

element.style["width"] = "calc(sin(30deg + 1.0471967rad ) * 100px)"
PASS 100 is 100

element.style["width"] = "calc(cos(30deg - 0.523599rad ) * 100px)"
PASS 100 is 100

element.style["width"] = "calc(sin(3.14159 / 2 + 1 - 1) * 100px)"
PASS 100 is 100

element.style["width"] = "calc(sin(100grad) * 100px)"
PASS 100 is 100

element.style["width"] = "calc(cos(0 / 2 + 1 - 1) * 100px)"
PASS 100 is 100

element.style["width"] = "calc(tan(30deg + 0.261799rad ) * 100px)"
PASS 100 is 100

element.style["width"] = "calc(tan(0.7853975rad ) * 100px)"
PASS 100 is 100

element.style["width"] = "calc(tan(3.14159 / 4 + 1 - 1) * 100px)"
PASS 100 is 100

element.style["width"] = "calc(sin(0.25turn) * 100px)"
PASS 100 is 100

element.style["width"] = "calc(cos(sin(cos(pi) + 1)) * 100px)"
PASS 100 is 100

element.style["width"] = "calc(sin(tan(pi/4)*pi/2) * 100px)"
PASS 100 is 100

element.style["width"] = "calc(sin(tan(pi/4)*pi/2) * 100px)"
PASS 100 is 100

element.style["width"] = "calc(sin(tan(pi/4)*pi/2) * 100px)"
PASS 100 is 100

element.style["min-width"] = "calc(sin(pi/2) * 100px)"
PASS 100 is 100

element.style["min-width"] = "calc(cos(pi - 3.14159265358979323846) * 100px)"
PASS 100 is 100

element.style["min-width"] = "calc(cos(e - 2.7182818284590452354) * 100px)"
PASS 100 is 100

element.style["min-width"] = "calc(sin(30deg + 1.0471967rad ) * 100px)"
PASS 100 is 100

element.style["min-width"] = "calc(cos(30deg - 0.523599rad ) * 100px)"
PASS 100 is 100

element.style["min-width"] = "calc(sin(3.14159 / 2 + 1 - 1) * 100px)"
PASS 100 is 100

element.style["min-width"] = "calc(sin(100grad) * 100px)"
PASS 100 is 100

element.style["min-width"] = "calc(cos(0 / 2 + 1 - 1) * 100px)"
PASS 100 is 100

element.style["min-width"] = "calc(tan(30deg + 0.261799rad ) * 100px)"
PASS 100 is 100

element.style["min-width"] = "calc(tan(0.7853975rad ) * 100px)"
PASS 100 is 100

element.style["min-width"] = "calc(tan(3.14159 / 4 + 1 - 1) * 100px)"
PASS 100 is 100

element.style["min-width"] = "calc(sin(0.25turn) * 100px)"
PASS 100 is 100

element.style["min-width"] = "calc(cos(sin(cos(pi) + 1)) * 100px)"
PASS 100 is 100

element.style["min-width"] = "calc(sin(tan(pi/4)*pi/2) * 100px)"
PASS 100 is 100

element.style["min-width"] = "calc(sin(tan(pi/4)*pi/2) * 100px)"
PASS 100 is 100

element.style["min-width"] = "calc(sin(tan(pi/4)*pi/2) * 100px)"
PASS 100 is 100
PASS successfullyParsed is true

TEST COMPLETE
